Tried to fix stat issues and proposal issues before working on UI stat bars
Change certain variable names to make it easier to understand.

Added a check for statChangeBus, though it is unfinished and not called yet.

Switched ActiveStatChange to rely on objects rather than scriptable objects for easy saving.

Added coroutine to disable selected buttons after period of time.

Added a proposalTitle for easier layout of proposal.